@@ -21,7 +21,7 @@
   ],
   "minimum-stability": "dev",
   "require": {
-    "php": "^7.1",
+    "php": "^7.2",
     "zendframework/zendframework": "^3.1"
   },
   "require-dev": {
