@@ -12,16 +12,55 @@ declare(strict_types=1);
 
 namespace General\Controller;
 
+use General\Controller\Plugin\GetFilter;
 use General\Entity\Currency;
 use General\Entity\ExchangeRate;
+use General\Service\FormService;
+use General\Service\GeneralService;
+use Zend\I18n\Translator\TranslatorInterface;
+use Zend\Mvc\Controller\AbstractActionController;
+use Zend\Mvc\Plugin\FlashMessenger\FlashMessenger;
 use Zend\View\Model\ViewModel;
 
 /**
  * Class CurrencyController
+ *
  * @package General\Controller
+ * @method GetFilter getFilter()
+ * @method FlashMessenger flashMessenger()
  */
-class ExchangeRateController extends GeneralAbstractController
+class ExchangeRateController extends AbstractActionController
 {
+    /**
+     * @var GeneralService
+     */
+    protected $generalService;
+    /**
+     * @var FormService
+     */
+    protected $formService;
+    /**
+     * @var TranslatorInterface
+     */
+    protected $translator;
+
+    /**
+     * ContentTypeController constructor.
+     *
+     * @param GeneralService      $generalService
+     * @param FormService         $formService
+     * @param TranslatorInterface $translator
+     */
+    public function __construct(
+        GeneralService $generalService,
+        FormService $formService,
+        TranslatorInterface $translator
+    ) {
+        $this->generalService = $generalService;
+        $this->formService = $formService;
+        $this->translator = $translator;
+    }
+
     /**
      * @return \Zend\Http\Response|ViewModel
      * @throws \Doctrine\ORM\ORMException
@@ -29,15 +68,16 @@ class ExchangeRateController extends GeneralAbstractController
      */
     public function newAction()
     {
-        $currency = $this->getGeneralService()->findEntityById(Currency::class, $this->params('currencyId'));
+        /** @var Currency $currency */
+        $currency = $this->generalService->find(Currency::class, (int)$this->params('currencyId'));
 
-        if (\is_null($currency)) {
+        if (null === $currency) {
             return $this->notFoundAction();
         }
 
         $data = $this->getRequest()->getPost()->toArray();
 
-        $form = $this->getFormService()->prepare(ExchangeRate::class, null, $data);
+        $form = $this->formService->prepare(ExchangeRate::class, $data);
         $form->remove('delete');
 
 
@@ -56,7 +96,15 @@ class ExchangeRateController extends GeneralAbstractController
                 $exchangeRate = $form->getData();
                 $exchangeRate->setCurrency($currency);
 
-                $result = $this->getGeneralService()->newEntity($exchangeRate);
+                $this->generalService->save($exchangeRate);
+
+                $this->flashMessenger()->setNamespace('info')
+                    ->addMessage(
+                        sprintf(
+                            $this->translator->translate("txt-exchangerate-for-%s-has-been-created-successfully"),
+                            $exchangeRate->getCurrency()->getName()
+                        )
+                    );
 
                 return $this->redirect()->toRoute(
                     'zfcadmin/currency/view',
@@ -67,10 +115,12 @@ class ExchangeRateController extends GeneralAbstractController
             }
         }
 
-        return new ViewModel([
-            'form'     => $form,
-            'currency' => $currency
-        ]);
+        return new ViewModel(
+            [
+                'form'     => $form,
+                'currency' => $currency
+            ]
+        );
     }
 
     /**
@@ -81,12 +131,14 @@ class ExchangeRateController extends GeneralAbstractController
     public function editAction()
     {
         /** @var ExchangeRate $exchangeRate */
-        $exchangeRate = $this->getGeneralService()->findEntityById(ExchangeRate::class, $this->params('id'));
+        $exchangeRate = $this->generalService->find(ExchangeRate::class, (int)$this->params('id'));
+
+        /** @var Currency $currency */
         $currency = $exchangeRate->getCurrency();
 
         $data = $this->getRequest()->getPost()->toArray();
 
-        $form = $this->getFormService()->prepare($exchangeRate, $exchangeRate, $data);
+        $form = $this->formService->prepare($exchangeRate, $data);
 
         if ($this->getRequest()->isPost()) {
             if (isset($data['cancel'])) {
@@ -99,7 +151,15 @@ class ExchangeRateController extends GeneralAbstractController
             }
 
             if (isset($data['delete'])) {
-                $this->getGeneralService()->removeEntity($exchangeRate);
+                $this->generalService->delete($exchangeRate);
+
+                $this->flashMessenger()->setNamespace('info')
+                    ->addMessage(
+                        sprintf(
+                            $this->translator->translate("txt-exchangerate-for-%s-has-been-deleted-successfully"),
+                            $exchangeRate->getCurrency()->getName()
+                        )
+                    );
 
                 return $this->redirect()->toRoute(
                     'zfcadmin/currency/view',
@@ -110,10 +170,18 @@ class ExchangeRateController extends GeneralAbstractController
             }
 
             if ($form->isValid()) {
-                /** @var Currency $exchangeRate */
+                /** @var ExchangeRate $exchangeRate */
                 $exchangeRate = $form->getData();
 
-                $this->getGeneralService()->updateEntity($exchangeRate);
+                $this->generalService->save($exchangeRate);
+
+                $this->flashMessenger()->setNamespace('info')
+                    ->addMessage(
+                        sprintf(
+                            $this->translator->translate("txt-exchangerate-for-%s-has-been-updated-successfully"),
+                            $exchangeRate->getCurrency()->getName()
+                        )
+                    );
 
                 return $this->redirect()->toRoute(
                     'zfcadmin/currency/view',
