@@ -19,25 +19,65 @@ namespace General\Controller;
 
 use Doctrine\ORM\Tools\Pagination\Paginator as ORMPaginator;
 use DoctrineORMModule\Paginator\Adapter\DoctrinePaginator as PaginatorAdapter;
+use General\Controller\Plugin\GetFilter;
 use General\Entity\Gender;
 use General\Form\GenderFilter;
+use General\Service\FormService;
+use General\Service\GeneralService;
+use Zend\I18n\Translator\TranslatorInterface;
+use Zend\Mvc\Controller\AbstractActionController;
+use Zend\Mvc\Plugin\FlashMessenger\FlashMessenger;
 use Zend\Paginator\Paginator;
 use Zend\View\Model\ViewModel;
 
 /**
+ * Class GenderController
  *
+ * @package General\Controller
+ * @method GetFilter getFilter()
+ * @method FlashMessenger flashMessenger()
  */
-class GenderController extends GeneralAbstractController
+class GenderController extends AbstractActionController
 {
     /**
-     * @return \Zend\View\Model\ViewModel
+     * @var GeneralService
      */
-    public function listAction()
+    protected $generalService;
+    /**
+     * @var FormService
+     */
+    protected $formService;
+    /**
+     * @var TranslatorInterface
+     */
+    protected $translator;
+
+    /**
+     * ContentTypeController constructor.
+     *
+     * @param GeneralService      $generalService
+     * @param FormService         $formService
+     * @param TranslatorInterface $translator
+     */
+    public function __construct(
+        GeneralService $generalService,
+        FormService $formService,
+        TranslatorInterface $translator
+    ) {
+        $this->generalService = $generalService;
+        $this->formService = $formService;
+        $this->translator = $translator;
+    }
+
+    /**
+     * @return ViewModel
+     */
+    public function listAction(): ViewModel
     {
         $page = $this->params()->fromRoute('page', 1);
-        $filterPlugin = $this->getGeneralFilter();
-        $contactQuery = $this->getGeneralService()
-            ->findEntitiesFiltered(Gender::class, $filterPlugin->getFilter());
+        $filterPlugin = $this->getFilter();
+        $contactQuery = $this->generalService
+            ->findFiltered(Gender::class, $filterPlugin->getFilter());
 
         $paginator = new Paginator(new PaginatorAdapter(new ORMPaginator($contactQuery, false)));
         $paginator::setDefaultItemCountPerPage(($page === 'all') ? PHP_INT_MAX : 20);
@@ -59,13 +99,13 @@ class GenderController extends GeneralAbstractController
     }
 
     /**
-     * @return array|ViewModel
+     * @return ViewModel
      */
-    public function viewAction()
+    public function viewAction(): ViewModel
     {
-        $gender = $this->getGeneralService()
-            ->findEntityById(Gender::class, $this->params('id'));
-        if (\is_null($gender)) {
+        $gender = $this->generalService->find(Gender::class, (int)$this->params('id'));
+
+        if (null === $gender) {
             return $this->notFoundAction();
         }
 
@@ -73,18 +113,15 @@ class GenderController extends GeneralAbstractController
     }
 
     /**
-     * Create a new template.
-     *
-     * @return \Zend\View\Model\ViewModel
+     * @return \Zend\Http\Response|ViewModel
+     * @throws \Doctrine\ORM\ORMException
+     * @throws \Doctrine\ORM\OptimisticLockException
      */
     public function newAction()
     {
-        $data = array_merge(
-            $this->getRequest()->getPost()->toArray(),
-            $this->getRequest()->getFiles()->toArray()
-        );
+        $data = $this->getRequest()->getPost()->toArray();
 
-        $form = $this->getFormService()->prepare(Gender::class, null, $data);
+        $form = $this->formService->prepare(Gender::class, $data);
         $form->remove('delete');
 
         if ($this->getRequest()->isPost()) {
@@ -96,8 +133,17 @@ class GenderController extends GeneralAbstractController
                 /* @var $gender Gender */
                 $gender = $form->getData();
 
-                $result = $this->getGeneralService()->newEntity($gender);
-                $this->redirect()->toRoute(
+                $result = $this->generalService->save($gender);
+
+                $this->flashMessenger()->setNamespace('info')
+                    ->addMessage(
+                        sprintf(
+                            $this->translator->translate("txt-gender-%s-has-been-created-successfully"),
+                            $gender->getName()
+                        )
+                    );
+
+                return $this->redirect()->toRoute(
                     'zfcadmin/gender/view',
                     [
                         'id' => $result->getId(),
@@ -111,39 +157,41 @@ class GenderController extends GeneralAbstractController
 
     /**
      * @return \Zend\Http\Response|ViewModel
+     * @throws \Doctrine\ORM\ORMException
+     * @throws \Doctrine\ORM\OptimisticLockException
      */
     public function editAction()
     {
-        $gender = $this->getGeneralService()
-            ->findEntityById(Gender::class, $this->params('id'));
+        /** @var Gender $gender */
+        $gender = $this->generalService->find(Gender::class, (int)$this->params('id'));
 
-        $data = array_merge(
-            $this->getRequest()->getPost()->toArray(),
-            $this->getRequest()->getFiles()->toArray()
-        );
+        $data = $this->getRequest()->getPost()->toArray();
 
-        $form = $this->getFormService()
-            ->prepare($gender, $gender, $data);
+        $form = $this->formService->prepare($gender, $data);
+        $form->remove('delete');
 
         if ($this->getRequest()->isPost()) {
             if (isset($data['cancel'])) {
                 return $this->redirect()->toRoute('zfcadmin/gender/list');
             }
 
-            if (isset($data['delete'])) {
-                $this->getGeneralService()->removeEntity($gender);
-
-                return $this->redirect()->toRoute('zfcadmin/gender/list');
-            }
-
             if ($form->isValid()) {
                 /** @var Gender $gender */
                 $gender = $form->getData();
-                $result = $this->getGeneralService()->updateEntity($gender);
-                $this->redirect()->toRoute(
+                $this->generalService->save($gender);
+
+                $this->flashMessenger()->setNamespace('info')
+                    ->addMessage(
+                        sprintf(
+                            $this->translator->translate("txt-gender-%s-has-been-removed-successfully"),
+                            $gender->getName()
+                        )
+                    );
+
+                return $this->redirect()->toRoute(
                     'zfcadmin/gender/view',
                     [
-                        'id' => $result->getId(),
+                        'id' => $gender->getId(),
                     ]
                 );
             }
