@@ -12,10 +12,16 @@ declare(strict_types=1);
 
 namespace General\Controller;
 
+use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\Tools\Pagination\Paginator as ORMPaginator;
 use DoctrineORMModule\Paginator\Adapter\DoctrinePaginator as PaginatorAdapter;
+use General\Controller\Plugin\GetFilter;
 use General\Entity\Log;
 use General\Form\EmailFilter;
+use General\Service\GeneralService;
+use Zend\I18n\Translator\TranslatorInterface;
+use Zend\Mvc\Controller\AbstractActionController;
+use Zend\Mvc\Plugin\FlashMessenger\FlashMessenger;
 use Zend\Paginator\Paginator;
 use Zend\View\Model\ViewModel;
 
@@ -23,31 +29,49 @@ use Zend\View\Model\ViewModel;
  * Class LogController
  *
  * @package General\Controller
+ * @method GetFilter getFilter()
+ * @method FlashMessenger flashMessenger()
  */
-class LogController extends GeneralAbstractController
+class LogController extends AbstractActionController
 {
+    /**
+     * @var GeneralService
+     */
+    protected $generalService;
+    /**
+     * @var EntityManager
+     */
+    protected $entityManager;
+    /**
+     * @var TranslatorInterface
+     */
+    protected $translator;
+
     /**
      * @return ViewModel
      */
-    public function listAction()
+    public function listAction(): ViewModel
     {
         $page = $this->params()->fromRoute('page', 1);
-        $filterPlugin = $this->getGeneralFilter();
-        $logQuery = $this->getGeneralService()
-            ->findEntitiesFiltered(Log::class, $filterPlugin->getFilter());
+        $filterPlugin = $this->getFilter();
+        $logQuery = $this->generalService
+            ->findFiltered(Log::class, $filterPlugin->getFilter());
 
         $paginator = new Paginator(new PaginatorAdapter(new ORMPaginator($logQuery, false)));
         $paginator::setDefaultItemCountPerPage(($page === 'all') ? PHP_INT_MAX : 20);
         $paginator->setCurrentPageNumber($page);
         $paginator->setPageRange(ceil($paginator->getTotalItemCount() / $paginator::getDefaultItemCountPerPage()));
 
-        $form = new EmailFilter($this->getEntityManager());
+        $form = new EmailFilter($this->entityManager);
         $form->setData(['filter' => $filterPlugin->getFilter()]);
 
-        if ($this->getRequest()->isGet()) {
-            if (null !== $this->getRequest()->getQuery('submit')) {
-                $this->getGeneralService()->truncateLog();
-            }
+        if ($this->getRequest()->isGet() && null !== $this->getRequest()->getQuery('submit')) {
+            $this->generalService->truncateLog();
+
+            $this->flashMessenger()->setNamespace('success')
+                ->addMessage(
+                    $this->translator->translate("txt-log-has-been-truncated-successfully")
+                );
         }
 
         return new ViewModel(
@@ -64,10 +88,9 @@ class LogController extends GeneralAbstractController
     /**
      * @return ViewModel
      */
-    public function viewAction()
+    public function viewAction(): ViewModel
     {
-        $log = $this->getGeneralService()
-            ->findEntityById(Log::class, $this->params('id'));
+        $log = $this->generalService->find(Log::class, (int)$this->params('id'));
         if (null === $log) {
             return $this->notFoundAction();
         }
