@@ -1,24 +1,22 @@
 <?php
 
 /**
- * Jield copyright message placeholder.
+ * Jield BV all rights reserved
  *
  * @category    Application
  *
- * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @author      Dr. ir. Johan van der Heide <info@jield.nl>
+ * @copyright   Copyright (c) 2004-2017 Jield BV (https://jield.nl)
  */
 
 declare(strict_types=1);
 
 namespace General\Controller\Plugin;
 
-use Interop\Container\ContainerInterface;
 use Zend\Http\Request;
+use Zend\Mvc\Application;
 use Zend\Mvc\Controller\Plugin\AbstractPlugin;
-use Zend\Mvc\Controller\PluginManager;
 use Zend\Router\Http\RouteMatch;
-use Zend\ServiceManager\ServiceLocatorInterface;
 
 /**
  * @category    Application
@@ -26,33 +24,61 @@ use Zend\ServiceManager\ServiceLocatorInterface;
 class GetFilter extends AbstractPlugin
 {
     /**
-     * @var PluginManager
+     * @var Request
      */
-    protected $serviceManager;
+    protected $request;
+    /**
+     * @var RouteMatch
+     */
+    protected $routeMatch;
     /**
      * @var array
      */
     protected $filter = [];
+    /**
+     * @var string|null
+     */
+    protected $order = 'name';
+    /**
+     * @var string|null
+     */
+    protected $direction = 'asc';
+    /**
+     * @var string|null
+     */
+    protected $query;
+
+    /**
+     * GetFilter constructor.
+     *
+     * @param Application $application
+     */
+    public function __construct(Application $application)
+    {
+        $this->routeMatch = $application->getMvcEvent()->getRouteMatch();
+        $this->request = $application->getMvcEvent()->getRequest();
+    }
+
 
     /**
      * Instantiate the filter
      *
      * @return GetFilter
      */
-    public function __invoke()
+    public function __invoke(): self
     {
-        $encodedFilter = urldecode((string)$this->getRouteMatch()->getParam('encodedFilter'));
+        $encodedFilter = urldecode((string)$this->routeMatch->getParam('encodedFilter'));
 
-        $order = $this->getRequest()->getQuery('order');
-        $direction = $this->getRequest()->getQuery('direction');
+        $order = $this->request->getQuery('order');
+        $direction = $this->request->getQuery('direction');
 
         //Take the filter from the URL
-        $filter = (array) json_decode(base64_decode($encodedFilter));
+        $filter = (array)json_decode(base64_decode($encodedFilter));
 
 
         //If the form is submitted, refresh the URL
-        if ($this->getRequest()->isGet() && !\is_null($this->getRequest()->getQuery('submit'))) {
-            $filter = $this->getRequest()->getQuery()->toArray()['filter'];
+        if ($this->request->isGet() && null !== $this->request->getQuery('submit')) {
+            $filter = $this->request->getQuery()->toArray()['filter'];
         }
 
         //Create a new filter if not set already
@@ -62,77 +88,23 @@ class GetFilter extends AbstractPlugin
 
         //Add a default order and direction if not known in the filter
         if (!isset($filter['order'])) {
-            $filter['order'] = 'name';
-            $filter['direction'] = 'asc';
+            $filter['order'] = 'id';
+            $filter['direction'] = 'desc';
         }
 
         //Overrule the order if set in the query
-        if (!\is_null($order)) {
+        if (null !== $order) {
             $filter['order'] = $order;
         }
 
         //Overrule the direction if set in the query
-        if (!\is_null($direction)) {
+        if (null !== $direction) {
             $filter['direction'] = $direction;
         }
 
         $this->filter = $filter;
 
         return $this;
-    }
-
-    /**
-     * @return RouteMatch
-     */
-    public function getRouteMatch(): RouteMatch
-    {
-        return $this->getServiceLocator()->get('application')->getMvcEvent()->getRouteMatch();
-    }
-
-    /**
-     * Get service locator.
-     *
-     * @return ServiceLocatorInterface
-     */
-    public function getServiceLocator(): ContainerInterface
-    {
-        return $this->serviceManager;
-    }
-
-    /**
-     * Proxy to the original request object to handle form
-     *
-     * @return Request
-     */
-    public function getRequest(): Request
-    {
-        return $this->getServiceLocator()->get('application')->getMvcEvent()->getRequest();
-    }
-
-    /**
-     * Return the filter
-     *
-     * @return array
-     */
-    public function getFilter(): array
-    {
-        return $this->filter;
-    }
-
-    /**
-     * @return string
-     */
-    public function getOrder(): string
-    {
-        return $this->filter['order'];
-    }
-
-    /**
-     * @return string
-     */
-    public function getDirection(): string
-    {
-        return $this->filter['direction'];
     }
 
     /**
@@ -146,12 +118,35 @@ class GetFilter extends AbstractPlugin
     }
 
     /**
-     * Set service locator.
-     *
-     * @param ServiceLocatorInterface $serviceLocator
+     * @return array
      */
-    public function setServiceLocator(ServiceLocatorInterface $serviceLocator)
+    public function getFilter(): array
     {
-        $this->serviceManager = $serviceLocator;
+        return $this->filter;
+    }
+
+
+    /**
+     * @return null|string
+     */
+    public function getOrder(): ?string
+    {
+        return $this->order;
+    }
+
+    /**
+     * @return null|string
+     */
+    public function getDirection(): ?string
+    {
+        return $this->direction;
+    }
+
+    /**
+     * @return null|string
+     */
+    public function getQuery(): ?string
+    {
+        return $this->query;
     }
 }
