@@ -19,24 +19,65 @@ namespace General\Controller;
 
 use Doctrine\ORM\Tools\Pagination\Paginator as ORMPaginator;
 use DoctrineORMModule\Paginator\Adapter\DoctrinePaginator as PaginatorAdapter;
+use General\Controller\Plugin\GetFilter;
 use General\Entity\Title;
 use General\Form\TitleFilter;
+use General\Service\FormService;
+use General\Service\GeneralService;
+use Zend\I18n\Translator\TranslatorInterface;
+use Zend\Mvc\Controller\AbstractActionController;
+use Zend\Mvc\Plugin\FlashMessenger\FlashMessenger;
 use Zend\Paginator\Paginator;
 use Zend\View\Model\ViewModel;
 
 /**
+ * Class TitleController
  *
+ * @package General\Controller
+ * @method GetFilter getFilter()
+ * @method FlashMessenger flashMessenger()
  */
-class TitleController extends GeneralAbstractController
+class TitleController extends AbstractActionController
 {
     /**
-     * @return \Zend\View\Model\ViewModel
+     * @var GeneralService
      */
-    public function listAction()
+    protected $generalService;
+    /**
+     * @var FormService
+     */
+    protected $formService;
+    /**
+     * @var TranslatorInterface
+     */
+    protected $translator;
+
+    /**
+     * TitleController constructor.
+     *
+     * @param GeneralService      $generalService
+     * @param FormService         $formService
+     * @param TranslatorInterface $translator
+     */
+    public function __construct(
+        GeneralService $generalService,
+        FormService $formService,
+        TranslatorInterface $translator
+    ) {
+        $this->generalService = $generalService;
+        $this->formService = $formService;
+        $this->translator = $translator;
+    }
+
+
+    /**
+     * @return ViewModel
+     */
+    public function listAction(): ViewModel
     {
         $page = $this->params()->fromRoute('page', 1);
-        $filterPlugin = $this->getGeneralFilter();
-        $query = $this->getGeneralService()->findEntitiesFiltered(Title::class, $filterPlugin->getFilter());
+        $filterPlugin = $this->getFilter();
+        $query = $this->generalService->findFiltered(Title::class, $filterPlugin->getFilter());
 
         $paginator
             = new Paginator(new PaginatorAdapter(new ORMPaginator($query, false)));
@@ -59,12 +100,12 @@ class TitleController extends GeneralAbstractController
     }
 
     /**
-     * @return \Zend\View\Model\ViewModel
+     * @return ViewModel
      */
-    public function viewAction()
+    public function viewAction(): ViewModel
     {
-        $title = $this->getGeneralService()->findEntityById(Title::class, $this->params('id'));
-        if (\is_null($title)) {
+        $title = $this->generalService->find(Title::class, (int)$this->params('id'));
+        if (null === $title) {
             return $this->notFoundAction();
         }
 
@@ -72,15 +113,15 @@ class TitleController extends GeneralAbstractController
     }
 
     /**
-     * Create a new template.
-     *
-     * @return \Zend\View\Model\ViewModel
+     * @return \Zend\Http\Response|ViewModel
+     * @throws \Doctrine\ORM\ORMException
+     * @throws \Doctrine\ORM\OptimisticLockException
      */
     public function newAction()
     {
-        $data = array_merge($this->getRequest()->getPost()->toArray(), $this->getRequest()->getFiles()->toArray());
+        $data = $this->getRequest()->getPost()->toArray();
 
-        $form = $this->getFormService()->prepare(Title::class, null, $data);
+        $form = $this->formService->prepare(Title::class, $data);
         $form->remove('delete');
 
 
@@ -93,11 +134,20 @@ class TitleController extends GeneralAbstractController
                 /* @var $title Title */
                 $title = $form->getData();
 
-                $result = $this->getGeneralService()->newEntity($title);
-                $this->redirect()->toRoute(
+                $this->generalService->save($title);
+
+                $this->flashMessenger()->setNamespace('info')
+                    ->addMessage(
+                        sprintf(
+                            $this->translator->translate("txt-title-%s-has-been-created-successfully"),
+                            $title->getName()
+                        )
+                    );
+
+                return $this->redirect()->toRoute(
                     'zfcadmin/title/view',
                     [
-                        'id' => $result->getId(),
+                        'id' => $title->getId(),
                     ]
                 );
             }
@@ -111,20 +161,16 @@ class TitleController extends GeneralAbstractController
      */
     public function editAction()
     {
-        $title = $this->getGeneralService()->findEntityById(Title::class, $this->params('id'));
+        /** @var Title $title */
+        $title = $this->generalService->find(Title::class, (int)$this->params('id'));
 
-        $data = array_merge($this->getRequest()->getPost()->toArray(), $this->getRequest()->getFiles()->toArray());
+        $data = $this->getRequest()->getPost()->toArray();
 
-        $form = $this->getFormService()->prepare($title, $title, $data);
+        $form = $this->formService->prepare($title, $data);
+        $form->remove('delete');
 
         if ($this->getRequest()->isPost()) {
             if (isset($data['cancel'])) {
-                return $this->redirect()->toRoute('zfcadmin/title/list');
-            }
-
-            if (isset($data['delete'])) {
-                $this->getGeneralService()->removeEntity($title);
-
                 return $this->redirect()->toRoute('zfcadmin/title/list');
             }
 
@@ -132,8 +178,17 @@ class TitleController extends GeneralAbstractController
                 /** @var Title $title */
                 $title = $form->getData();
 
-                $title = $this->getGeneralService()->updateEntity($title);
-                $this->redirect()->toRoute(
+                $this->generalService->save($title);
+
+                $this->flashMessenger()->setNamespace('info')
+                    ->addMessage(
+                        sprintf(
+                            $this->translator->translate("txt-title-%s-has-been-updated-successfully"),
+                            $title->getName()
+                        )
+                    );
+
+                return $this->redirect()->toRoute(
                     'zfcadmin/title/view',
                     [
                         'id' => $title->getId(),
