@@ -2,52 +2,41 @@
 /**
  * ITEA Office all rights reserved
  *
- * PHP Version 7
+ * @category  Content
  *
- * @category    General
+ * @author    Johan van der Heide <johan.van.der.heide@itea3.org>
+ * @copyright Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
+ * @license   https://itea3.org/license.txt proprietary
  *
- * @author      Johan van der Heide <johan.van.der.heide@itea3.org>
- * @copyright   Copyright (c) 2004-2017 ITEA Office (https://itea3.org)
- * @license     https://itea3.org/license.txt proprietary
- *
- * @link        http://github.com/iteaoffice/main for the canonical source repository
+ * @link      https://itea3.org
  */
+
 declare(strict_types=1);
 
 namespace General\Factory;
 
 use Doctrine\ORM\EntityManager;
 use General\Service\FormService;
-use General\Service\GeneralService;
 use Interop\Container\ContainerInterface;
 use Zend\ServiceManager\Factory\FactoryInterface;
 
 /**
  * Class FormServiceFactory
  *
- * @package General\Factory
+ * @package Admin\Factory
  */
 final class FormServiceFactory implements FactoryInterface
 {
     /**
      * @param ContainerInterface $container
-     * @param string $requestedName
-     * @param array|null|null $options
+     * @param string             $requestedName
+     * @param array|null         $options
      *
-     * @return mixed
+     * @return FormService
      */
     public function __invoke(ContainerInterface $container, $requestedName, array $options = null): FormService
     {
         /** @var FormService $formService */
-        $formService = new $requestedName();
-        $formService->setServiceLocator($container);
-        /** @var EntityManager $entityManager */
-        $entityManager = $container->get(EntityManager::class);
-        $formService->setEntityManager($entityManager);
-        /** @var GeneralService $generalService */
-        $generalService = $container->get(GeneralService::class);
-        $formService->setGeneralService($generalService);
-
-        return $formService;
+        return new $requestedName($container, $container->get(EntityManager::class));
     }
 }
